# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: SZL Holdings API
  version: 1.0.0
  description: Unified API for SZL Holdings platforms - Beacon, Nimbus, Zeus, INCA AI, Dream Era
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zeus/modules:
    get:
      operationId: listZeusModules
      tags: [zeus]
      summary: List system modules
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZeusModule"
    post:
      operationId: createZeusModule
      tags: [zeus]
      summary: Create a module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZeusModule"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZeusModule"
  /zeus/modules/{id}:
    put:
      operationId: updateZeusModule
      tags: [zeus]
      summary: Update a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZeusModule"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZeusModule"
    delete:
      operationId: deleteZeusModule
      tags: [zeus]
      summary: Delete a module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /zeus/logs:
    get:
      operationId: listZeusLogs
      tags: [zeus]
      summary: List system logs
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZeusLog"
    post:
      operationId: createZeusLog
      tags: [zeus]
      summary: Create a log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZeusLog"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZeusLog"

  # ---- INCA AI ----
...
components:
  schemas:
...
    ZeusModule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [active, inactive, error, updating]
        category:
          type: string
        uptime:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, version, status, category, uptime, createdAt]
    CreateZeusModule:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [active, inactive, error, updating]
        category:
          type: string
        uptime:
          type: number
      required: [name, description, version, status, category, uptime]
    ZeusLog:
      type: object
      properties:
        id:
          type: integer
        level:
          type: string
          enum: [info, warn, error, debug]
        message:
          type: string
        module:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, level, message, module, createdAt]
    CreateZeusLog:
      type: object
      properties:
        level:
          type: string
          enum: [info, warn, error, debug]
        message:
          type: string
        module:
          type: string
      required: [level, message, module]

    # INCA AI
...