# lib/api-zod/src/generated/types/createZeusLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified API for SZL Holdings platforms - Beacon, Nimbus, Zeus, INCA AI, Dream Era
 * OpenAPI spec version: 1.0.0
 */
import type { CreateZeusLogLevel } from "./createZeusLogLevel";

export interface CreateZeusLog {
  level: CreateZeusLogLevel;
  message: string;
  module: string;
}